Move alloc_extend_naive to the NPU allocator
Relocate the large NPU extend fallback without changing its signature, body, branch, or call behavior. Remove the obsolete main-tree/package exposure and use the concrete paged allocator import to keep the cut acyclic.

Extraction: awk '/^def alloc_extend_naive\(/ {copying=1} copying && /^class PagedTokenToKVPoolAllocator/ {exit} copying {print}' python/sglang/srt/mem_cache/allocator/paged.py > /private/tmp/op30-wave1-fe18daa8-alloc_extend_naive.before.py

Patch: apply_patch < /private/tmp/op30-wave1-fe18daa8-mechanical.patch

Verification: cmp /private/tmp/op30-wave1-fe18daa8-alloc_extend_naive.before.py /private/tmp/op30-wave1-fe18daa8-alloc_extend_naive.after.py
