+++ python/sglang/srt/mem_cache/allocator/__init__.py
@@ -3,7 +3,6 @@
 from sglang.srt.mem_cache.allocator.base import BaseTokenToKVPoolAllocator
 from sglang.srt.mem_cache.allocator.paged import (
     PagedTokenToKVPoolAllocator,
-    alloc_extend_naive,
 )
 from sglang.srt.mem_cache.allocator.token import TokenToKVPoolAllocator
 
@@ -11,5 +10,4 @@ __all__ = [
     "BaseTokenToKVPoolAllocator",
     "PagedTokenToKVPoolAllocator",
     "TokenToKVPoolAllocator",
-    "alloc_extend_naive",
 ]
